feat: implementado Single Page Application e separação de arquivos para melhor organização
