+++ index.html
@@ -31,7 +31,4 @@
 
   </body>
   <script src="main.js" type="module"></script>
-
-  <script src="modules/sorteadorNumeros.js"></script>
-
 </html>
